change placeholder service
It's not the first time it happens. `placehold.it` is now `placeholder.com`. Still it's not the first time they let the certificates expire. 

It's certificate is expired since 25th of december this year. 

I am switching here for a new placeholder service, but still, I think we might want to have our own image creation process (maybe just css will be enough ? )

Best,
Pedro